Refactor following review points
