+++ README.md
@@ -1,7 +1,7 @@
 # Bookstore App
 
 
-- This is React Bookstore app which was built fully using React and JSX. In this app, you can manage bookstore assets in a very easily and fast.
+- This is React Bookstore app which was built fully using React and JSX. In this app, you can manage bookstore assets very easily and fast.
 
 ## Live Demo
