+++ src/components/App.js
@@ -1,5 +1,5 @@
-import BooksList from './components/BooksList';
-import BooksForm from './components/BooksForm';
+import BooksList from './BooksList';
+import BooksForm from './BooksForm';
 
 function App() {
   return (
