+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import store from './reducers/index';
-import App from './App';
+import App from './components/App';
 
 ReactDOM.render(
   <React.StrictMode>
